Vfw4 - a standalone Win32 API exercise
This is a sample Win32 application written from scratch to solidify
learning C++ from the textbook, "C++ Primer, 4th Edition" by Lippman,
LaJoie and Moo.

The app builds in 32 and 64 bit versions, and runs on Win7, 8 and 10. It
exercise the VFW (Video for Windows) API, Microsoft's first Windows
Multimedia API. VFW is deprecated, but still ships in Windows and runs
fine. To my knowledge, this is the only 64 bit VFW application in
existance.